backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# OpenAI and LLM
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.2
langchain-community==0.0.16

# RAG and Embeddings
sentence-transformers==2.3.1
faiss-cpu==1.7.4
tiktoken==0.5.2

# Graph Database
neo4j==5.16.0

# Data Processing
numpy==1.26.3
pandas==2.1.4

# HTTP and CORS
httpx==0.26.0
python-dotenv==1.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
